+++ pyhidra/gui.py
@@ -1,8 +1,32 @@
+import contextlib
 import os
+import platform
 import sys
 import warnings
 
 from pyhidra import get_current_interpreter as _get_current_interpreter
+
+
+def _gui():
+    if platform.system() == 'Windows':
+        # gui_script works like it is supposed to on windows
+        gui()
+        return
+
+    pid = os.fork()
+    if pid != 0:
+        # original process can exit
+        return
+
+    # close stdin, stdout and stderr so the jvm can't use the terminal
+    # these must be closed using os.close and not sys.stdout.close()
+    os.close(sys.stdin.fileno())
+    os.close(sys.stdout.fileno())
+    os.close(sys.stderr.fileno())
+
+    # run the application
+    gui()
+
 
 def gui():
     """