+++ pyhidra/java/plugin/__init__.py
@@ -1,1 +1,4 @@
+"""
+Internal use only
+"""
 from .handler import install