+++ tests/test_core.py
@@ -1,91 +1,54 @@
 
-import pathlib
+from pathlib import Path
 import textwrap
 import importlib
-import jpype
 import sys
+import jpype
 import pyhidra
-from pyhidra.__main__ import _get_parser, PyhidraArgs
-from pyhidra.script import PyGhidraScript
 
-#pylint: disable=protected-access, missing-function-docstring
-
-# hack fix so capsys works correctly
-PyGhidraScript._print_wrapper = lambda self: print
+EXE_NAME = "strings.exe"
 
 
-def test_run_script(capsys, strings_exe):
-    script_path = pathlib.Path(__file__).parent / "example_script.py"
-
+def test_run_script(capsys, shared_datadir: Path):
+    strings_exe = shared_datadir / EXE_NAME
+    script_path = shared_datadir / "example_script.py"
     pyhidra.run_script(strings_exe, script_path, script_args=["my", "--commands"])
     captured = capsys.readouterr()
 
     expected = textwrap.dedent(f"""\
         {script_path} my --commands
-        strings.exe - .ProgramDB
+        my --commands
+        {EXE_NAME} - .ProgramDB
     """)
 
     assert captured.out == expected
 
 
-def test_open_program(strings_exe):
+def test_open_program(shared_datadir: Path):
+    strings_exe = shared_datadir / EXE_NAME
     with pyhidra.open_program(strings_exe, analyze=False) as flat_api:
-        assert flat_api.currentProgram.name == "strings.exe"
+        assert flat_api.currentProgram.name == strings_exe.name
         assert flat_api.getCurrentProgram().listing
         assert flat_api.getCurrentProgram().changeable
 
 
-def test_no_project(capsys):
-    script_path = pathlib.Path(__file__).parent / "projectless_script.py"
-
-    pyhidra.run_script(None, script_path)
+def test_no_project(capsys, shared_datadir: Path):
+    pyhidra.run_script(None, shared_datadir / "projectless_script.py")
     captured = capsys.readouterr()
     assert captured.out.rstrip() == "projectless_script executed successfully"
 
 
-def test_no_program(capsys):
-    script_path = pathlib.Path(__file__).parent / "programless_script.py"
-    project_path = pathlib.Path(__file__).parent / "programless_ghidra"
+def test_no_program(capsys, shared_datadir: Path):
+    script_path = shared_datadir / "programless_script.py"
+    project_path = shared_datadir / "programless_ghidra"
 
     pyhidra.run_script(None, script_path, project_path, "programless")
     captured = capsys.readouterr()
     assert captured.out.rstrip() == "programless_script executed successfully"
 
 
-def test_arg_parser(strings_exe):
-    script_path = pathlib.Path(__file__).parent / "example_script.py"
-    parser = _get_parser()
-    strings_exe = str(strings_exe)
-    args = [str(script_path), strings_exe]
-    args1 = PyhidraArgs()
-    args2 = PyhidraArgs()
-    parser.parse_args(args, namespace=args1)
-    args.reverse()
-    parser.parse_args(args, namespace=args2)
-    assert args1 == args2
-    args.insert(0, "-v")
-    args1 = parser.parse_args(args, namespace=PyhidraArgs())
-    assert args1.verbose is True
-    args = ["--project-name", "stub_name"] + args
-    args1 = parser.parse_args(args, namespace=PyhidraArgs())
-    assert args1.project_name == "stub_name"
-    args = ["--project-path", str(script_path.parent)] + args
-    args1 = parser.parse_args(args, namespace=PyhidraArgs())
-    assert args1.project_path == script_path.parent
-
-    # two scripts are ok
-    parser.parse_args([str(script_path), str(script_path)], namespace=PyhidraArgs())
-
-    try:
-        # two binary files without a script is not
-        parser.parse_args([strings_exe, strings_exe], namespace=PyhidraArgs())
-        assert False
-    except ValueError:
-        pass
-
-
-def test_import_script(capsys):
-    script_path = pathlib.Path(__file__).parent / "import_test_script.py"
+def test_import_script(capsys, shared_datadir: Path):
+    script_path = shared_datadir / "import_test_script.py"
     pyhidra.run_script(None, script_path)
     captured = capsys.readouterr()
     assert captured.out.rstrip() == "imported successfully"
@@ -155,7 +118,5 @@ def test_import_ghidra_base_java_packages():
     # Test _Jpackage handles import that doesn't exist
     try:
         import pdb_.doesntexist
-    except ImportError as e:
+    except ImportError:
         pass
-
-